requirements.txt
aioredis==1.3.1
alabaster==0.7.12
asgi-redis==1.4.3
asgiref==3.2.2
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
autopep8==1.4.4
Beaker==1.11.0
certifi==2019.11.28
cffi==1.13.2
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
Django==2.2.7
django-appconf==1.0.3
django-autoslug-iplweb==1.9.4.dev0
django-cities-light==3.5.0
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-crispy-forms==1.9.1
django-crontab==0.7.1
django-js-asset==1.2.2
django-select2==7.2.3
djangorestframework==3.11.0
docopt==0.6.2
ecdsa==0.14.1
future==0.18.2
hiredis==1.0.1
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
msgpack-python==0.5.6
mysqlclient==1.4.6
num2words==0.5.10
numpy==1.18.0
oic==0.13.0
pandas==0.25.3
pdfkit==0.6.1
Pillow==6.2.1
pkg-resources==0.0.0
progressbar2==3.47.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
pycryptodomex==3.9.4
PyHamcrest==1.9.0
pyjwkest==1.4.2
pylint==2.5.3
pyOpenSSL==19.1.0
PyPDF2==1.26.0
python-dateutil==2.8.1
python-utils==2.3.0
pytz==2019.3
razorpay==1.2.0
redis==2.10.6
reportlab==3.5.32
requests==2.22.0
rsa==4.0
service-identity==18.1.0
six==1.13.0
sqlparse==0.3.0
toml==0.10.1
Twisted==19.10.0
txaio==18.8.1
typed-ast==1.4.1
Unidecode==1.1.1
urllib3==1.25.7
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.3
zope.interface==4.7.1
